custom_mcp_tools/requirements.txt
mcp-python==0.2.4
openai>=1.4.0
pinecone-client>=3.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0 